requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
openai==1.51.0
supabase==2.9.0
PyJWT==2.9.0
cryptography==43.0.1
httpx==0.27.2
notion-client==2.2.1
slack-sdk==3.33.0
langgraph==0.2.38
requests==2.32.3
python-multipart==0.0.12
